Newline escapes are moved to the next line for more precise control
Instead of
	let x = foo \
	bar

We want to make it possible to differentiate between:

	let x = foo
	\bar
	let x = foo
	\ bar

As spaces are sometimes more significant than usual in N, it makes sense
to move the escape character to the next line. And we've always
advocated for operators to be on the next line when breaking up a long
expression.

It's a little similar to the practice in some programming languages that
don't allow a trailing comma of placing the comma in a list at the
beginning of the next line (to facilitate the appending of new entries).