bootstrap:types: typ_hash(): ensure typ is not tentative
The 'hash' in a tentative typ is maintained as the typ gets linked to a
new typs. So 'hash' can be used in typ_equal(), etc. But it cannot be
used in a hash table *IF* elements of the table are re-linked while the
table exists. typset can be used in typ_isalist_foreach() with tentative
types when typs are not actively being linked. E.g. unify_generics()
typ_isalist_each(find_instance_of) is fine. So we cannot
assert(!typ_is_tentative(*a)) in general.