5/10: bootstrap:types: typ_hash_ready(): whether a hash is available
Not only do typ not have a well-defined hash early in the compilation
(before typ_hash_ready() is true), typ are not immutable, and their hash
can change (tentative types, tentative generics).

Therefore, while it's OK to store in a typset a typ with a non-final
hash, one must take care to not store a typ in a typset across hash
changes, or further lookup of that typ element will fail (it would have
been necessary to re-hash the element in the typset when the hash
changes, say).

This is a very difficult property to enforce, however (typ_hash_ready()
does not help with the update of hashes, only with their
initialization). It would be nice to get rid of it.