8/10: bootstrap:types: typ_is_concrete()
Examples of non-concrete types: uninstantiated generics or instances
within an instantiated generic which are instantiated over non-tentative
interfaces.

We will likely have to reconsider the latter case. (foo `t) is concrete
if `t is only ever used in code in a dyn (a dyn of any non-tentative
intf is concrete). We don't know how to check that property right now.
Instead we say that you have to use (foo @`t). This may not be
sufficient in general. This breaks `environment as written, although we
may be able to fix that independently.