Use '..' for range instead of '...'
Swift is using .. for (opened) range and ... for closed range. While we're
not convinced of the wisdom of introducing such an typo prone syntax by
supporting both, we see no reason to use a different syntax for (opened)
ranges. Many languages already use '..'. As long as there was no
language using '...' it was fine to be different; not anymore.