bootstrap: Switch to uppercase first-letter for export
The naming convention remains lowercase words separated by underscore.
But for exported symbols, the first letter is uppercase.

So it's Html_output; Html_document and Write_html_document; Tcp_window;
String and To_string; Float and round_even_float; etc.

We've gone back and forth on this before, but we resolutely like that it
becomes explicit, at the use-site, whether the symbol is public or
private.