bootstrap: type the intf portion of a fun/met member name declaration
Initially, we intended to make mandatory to always declare the intf of
origin for implemented members, i.e.

	struct a =
		isa `Show

	a met `Show.Show buf:@!`String_buffer = void

To make this possible, we now type the '`Show' portion of the name, for
the benefit of step_check_exhaustive_intf_impl(). That check is still
missing though.

However, we're going to abandon the systematic requirement for
specifying the intf. It quickly becomes painful and is not that helpful
(unlike in languages with inheritance, with an 'override' directive,
these method declarations are not replacing/hiding an existing
behavior).

Instead, the intent is to require such specification when implementing
several members from multiple interfaces that have the same name. This
is not yet supported either.