bootstrap:types: typ_isa(): enable the use of quickisa
The cost of typ_isa() was significant, not anymore.

It is important to note that there are likely simpler solution to the
cost of a recursive typ_isa():
 - (a) quickisa: precompute common answers in typset stored in each typ;
 - (b) caching: store the result of typ_isa() for quick re-use;
 - (c) special case: most of the typ_isa() calls, at least on the
   current testing code-base, are related to tests like
	(Ref U8) isa (`Any_ref `Any)
   (note the `Any on the RHS). We would likely be able to speed up
   a naive typ_isa() by simply hand-coding these cases.

A combination of options (b) and (c) would likely be enough (and maybe
simpler than quickisa). An alternative to consider for ncc.