1/2: bootstrap: rely on passsem to enforce nullable
We use to have:

	intf `Any_nullable_ref t:`Any =
	       isa `Any_any_ref `Trivial_ctor
	intf `Any_nullable_mutable_ref t:`Any =
	       isa (`Any_nullable_ref t)

We now have:

	intf `Any_nullable_ref t:`Any =
	       isa (`Any_ref t) `Trivial_ctor
	intf `Any_nullable_mutable_ref t:`Any =
	       isa (`Any_nullable_ref t) (`Any_mutable_ref t)

As far as typing and unification are concerned, nullable references and
non-nullable references are now compatible. The semantic pass enforces
that nullable references are known to be nonnull before they are used.

What we had before was a bit of an hybrid. In some places, it was
necessary to use a Nonnull_cast, even after checking the reference
against null. It didn't really make sense.