bootstrap:instantiation: fast lookup for suitable existing instances
The linear search through existing instances was quite expensive. This
change saves 20% of total runtime.

As usual (see typset, topdeps), we must be careful when storing
tentative typs in a lookup data structure.

We also change when tentative typs are finalized (link_finalize()
and schedule_finalization()), instead of using
step_gather_final_instantiations(). We still need a
step_gather_remaining_weakly_concrete() as, by definition, weakly
concrete types that are not linked to something else by the end of
bodypass0 finalize to themselves.

By finalizing tentative typs earlier than before (at the end of a
unify*() invokation, as opposed to the end of bodypass0 in
step_gather_final_instantiations()), we can get rid of link_members()
too, and more generally feel better about all of this.