bootstrap: complete support for ov ops; restrict to native unsigned; allow bw ops on all ints
We can be a little bit more discriminate about what we allow and what we
do not allow. >> is always well-defined on integers, signed or unsigned,
even when the bitwidth is unknown. So are bw& bw| and bw^.

However, << is renamed ov<< and is no more allowed on signed integers, as
we can loose the sign.

`Overflow_arithmetic in general is not allowed on signed integers
anymore.

The goal is to minimize ill-defined or often-misunderstood behaviors
(the latter is not much better in practice), in application of the
principle of "benign-looking code should be just that". This API may
still change.