bootstrap:phi: we choose not to track WITHIN (for now)
With the kind of constraints supported at this point, little would be
gained from it.

Additionally, as they refer to a global environment, it is impossible to
perform any analysis between the different functions that use the
environment. Analysis can only be performed locally, in a single
function. It is not entirely clear how this works.