bootstrap: revert pre,post,invariant,example syntax to accepting a block
We're back to:
	example
		a == b

pre,post,invariant are now going to be primitives, tightly linked to the
semantic checker, so it doesn't really make sense anymore to have a
function-like syntax.

'example' is likely going to get additional features which will make it
look more like a function definition than a function call.