lib/n/builtins: rename 'Unsafe_{,un}signed' to Signed and Unsigned
These functions can assert, but that doesn't make them unsafe.
Especially because with the use of pre (semantics), we should be able to
often bypass the assert.

Unsafe_* would mean that there is no check at all. We don't actually
have exported methods for that yet. Not sure what their name should be.