bootstrap: overlay types: get rid of weakly concrete support
We're removing the ability to transparently convert boolean and string
literals to other types. We're keeping this capability for integers and
reintroducing it for array literals.

For array literals, we now have a (`__literal_slice__ `Any) that
converts down to a concrete type on first use -- whether when used in a
TBI_ARRAY_CTOR context (linking to the context type), or at the first
binary accessor on it (linking it to a Slice).