bootstrap: 3/3: overlay typs
For tentative generic instantiation, our approach was to create a
deepcopy of the generic functor, do all the passes until the last
passfwd, and use the typ of that partially-instantiated definition.

But this approach was slow, especially because of all the interfaces to
instantiate from the isalist, and because references are not built in
the compiler, but defined as regular types in builtins.n, so they had to
be instantiated as well.

With the new approach, instantiating a generic is merely creating a
wrapper around its typ that knows how to rewrite on the fly the generic
arguments of the functor to their instantiated versions. This required a
deep refactoring to guarantee that the typs coming from the
uninstantiated generic functor are always rewritten through the overlay
(hence all the new typ_* and tit_* functions).

This new approach is significantly faster. Concrete generic
instantiations still follow the original approach, as we need to perform
the bodypasses on the body of concrete instances. While the body of the
generic functor is also bodypassed, the different generic arguments can
lead to different transformations of the body (esp. in bodypass1), so we
cannot just delay the body rewriting to, say, cprinter.