bootstrap:cprinter: stop distinguishing const/non-const refs in C
ncc0 knows about reference compatibility between const and
non-const (e.g. return a non-const ref in a function returning a const
ref). So does the C compiler.

But ncc0 also knows about this rule for tuples of references. But the C
compiler doesn't know about it. So we would have to introduce explicit
conversion operations in the generated C code. This is painful work and
may be somewhat costly.

Instead, we represent all references the same way in the generated C
code (non-const refs). This is fine, as ncc0 is advanced enough that we
feel more comfortable not having the C compiler double-check our work,
but we do loose constness in function arguments, which could limit the C
compiler's ability to optimize.

TODO, maybe: Try and restore constness for function arguments in the
generated C code.