bootstrap: change example to take a Void block, not a Bool
While it was cute to have
	example true != false

On most real-life examples, multiple conditions are checked. So assert
ended up being used multiple times. The last statement had a different
status and that was a little strange. Often, that last statement had to
be a plain 'true'.

The alternative of making example blocks special, so that any statement
can either be Void, or Bool (in which case it would asserted over), was
rejected as being too fancy for its own good.