lib/n/bytes: return (slice, Error) not (Error, slice)
By convention (to be enforced by semantic constraints), when a function
returns a tuple that contains an Error value, if the error is not OK,
then the values that appear after the Error are to be ignored.

However, for Read and Write, the slice is always valid. On error, the
slice may be empty, or it may contain some valid content on EOF.