bootstrap: vararg: support passing a slice as vararg
The call syntax is
	foo a opt=1 ...v
	foo a opt=1 ...{1 2 3}

This still compiles to code that uses C vararg. However, when passing a
slice in, the _$Nvarargcount argument is set to -1 and a single argument
is passed after the count: a pointer to the slice.

The Vararg.Next builtin either uses va_arg or gets the element from the
slice, depending on the value of the _$Nvarargcount argument.