bootstrap: add pretend support for 'var': like 'let' but allows mutation
We held out introducing 'var' for as long as we could.
'let ... such ...' is great, but despite our best efforts, there really
are cases where adding a level of indentation for the such block is just
a pain. We suspected things may come to this. However, we have acquired
a strong taste for 'let ... such ...' and we hope other can be convinced
to give it a preferred status to a plain 'var'.

This commit only adds support in the lexer: 'var' is parsed like 'let'.
This works because the compiler doesn't yet enforce constness of
variables declared with 'let'.

TODO: complete support.