bootstrap:cprinter: call through dyntable directly
Do not generate "defintf members" stubs. This is simpler, faster (a
bit), generates less C code.

Also, this change was necessary to properly call methods defined on a
"parent" of the intf of the dyn itself. The method prototypes are not
copied into the descendant intf, but referenced directly from the
parent. In the codegen, we were blindly getting the prototype from the
parent. But we were also reusing the "defintf member" stub generated *for*
the parent intf, but they have different dyntables! We had an entirely
bogus cast in the dyn call code before.