bootstrap: introduce syntax for creating union variant from external payload
	union un =
		| A -> I32

For initialization, the following forms are supported:
	let x = {}:un.A
	let x = 1:un.A
	let x = y:un.A

When accessing a variant with an external payload, the expression 'x.A'
is typed with the external payload type 'I32' directly.

We still do not have syntax to obtain the type of a variant payload.
(variant un A) is the tentative syntax, but if we're keeping
hierarchical unions, we would have to somehow support that case as well.