bootstrap:topdeps: fix subtle missing "inline struct" ordering constraint
When a type from module mod1 depends on a generic defined in module
mod2, that itself "inline struct" depends on a non-generic also in module
mod2.

E.g. mod1.foo depends on (n.builtins.Vararg U8), which depends on
n.builtins.Varargint. The definition for (n.builtins.Vararg U8) may be
written out in mod1.n.o.h, meaning the full definition of
n.builtins.Varargint must be visibile by that point, i.e. mod1 must have
an inline struct dependency on n.builtins (it must be an inline import).

We were previously not adding the inline import for that case, and would
fail unless the dependency was otherwise necessary.