bootstrap: support for relative imports
Normal imports are looked up in a well-defined list of prefixes
(currently $PWD and $PWD/lib).

Until now, we also looked in the directory where the importing module
was found. Now, a relative import must be used (and it only looks in
that directory, not in the global list of prefixes).

	import .relative
	import .a.b