bootstrap:import: change import naming: 'import a.b.c' imports under 'c.Name'
This is a much nicer rule than previously, when we created local,
complete names 'a.b.c.Name'. Modules names are great to avoid repetition
in naming, but we typically don't need the whole module path. If
ambiguities arise, we can use a 'import a.b.c as d' form (TODO: add
support for it).