bootstrap:lir: lift restriction on except on DEFINC expr
The current DEFINCOMPLETE resolution, which is unable to handle method
calls on incomplete types, prevented the use of this form on bare Error
global constants such as OK or INVAL:
	except = INVAL -- won't work
	except = Error.INVAL -- works

Indeed, the LIR transform created (roughly):
	LET
	 DEFNAME
	  IDENT(tmp):definc
	  IDENT(INVAL)
	BLOCK
	 IF
	  CALL
	   BIN
	    IDENT(tmp):definc
	    IDENT(Operator_test) -- BOOM!
	 BLOCK
	  THROW
	   IDENT(tmp) -- only now would it unify to Error
	 BLOCK
	  NOOP

For now, we choose to resolve this by constraining expr to Error when
'except' is found, leading to:
	  CALL
	   BIN
	    TYPECONSTRAINT
 	     IDENT(tmp)
	     DIRECTDEF -- to Error
	    IDENT(Operator_test)

This prevents reuse of try/catch/except/throw/... with non-Error types.
This is something we may want to allow in the future, at which point we
would hopefully have addressed the DEFINC limitation.