bootstrap:autointf: `Inherit support
`Inherit is used to support a conditional 'isa' declaration, typically
depending on the isalist of a generic argument. This is a bit of a
tentative feature. It will be used by Optional in the upcoming commit.
For this limited use-case, we could have special handling in autointf
for Optional.

To be really interesting, we would need support for conditional member
declaration: "if t isa `Copyable, then define Slicebuf met Push"...