bootstrap:types: typ_isa(): remove unnecessary code for dyn handling
typ_isa(@i, intf) would check whether i is an intf (meaning @i is a
dyn), and if so, return typ_isa(i, intf).

This hasn't been necessary for a long time, and is in fact wrong in many
cases where typ_isa() gets used.

This case is now handled by special case in unify.c.