n/builtins: convert functions: all-purpose integer conversions (useful in generics)
Our naming conventions for integer conversion operations are meant to be
explicit in what they do: whether they are safe or may assert on range
error.

A generic such as
	(fun t:`Integer u:`Integer) add x:t y:u = t
		return x + y
cannot be easily defined in terms of conversion operations (one cannot
simply do a cast to (t) on both operands, like one would in C++). The
methods to use depend on t and u.

	(fun t:`Integer u:`Integer) add x:t y:u = t
		return x + (Convert t) y
	(fun t:`Integer u:`Integer) add x:t y:u = t
		return x + (Force_convert t) y

are two possible definitions using the convert family of functions. They
have different semantics: the first will assert on range error. The
second will reinterpret and extend/trim the bit representation.

Try_convert returns an error on range error.

Safe_convert is only available for type pairs that can safely be converted.
	Safe_convert I16 U8 -- is available
	Safe_convert I8 U8 -- is not available
The second form results in a compilation failure.

Only Force_convert is implemented yet.