bootstrap:cprinter: stop casting numerical constants to their concrete type
We thought the approach of converting
	let x:I32 = 1 * 2 + 3 / 4
into
	let x:I32 = 1:I32 * 2:I32 + 3:I32 / 4:I32
was sensible and would make it easy to track overflow/underflow, etc.

But in fact, this turns out to be entirely misguided.

A case such as
	fun foo = I8
		return -0x80
which gets effectively transformed into
	fun foo = I8
		return -(0x80:I8)
illustrates how naive the idea is. 0x80:I8 overflows... Even though the
function definition is perfectly fine.

Instead of fixing this particular case, we're coming to the conclusion
that Go probably got it right all along: perform all numerical
computations on constants with unlimited precision, then convert to the
final concrete type, checking for overflow only then. We will probably
adopt a similar strategy down the road. TODO

In the meantime, all integer constants are printed with the postfix LL
and no type conversion. This isn't ideal either, but works in more cases
and is more like C.