bootstrap:parser: try/catch: do not consume last EOL after last catch
Indeed, it is part of the try statement as a whole (as catch are not
nested deeper than the try, and we treat the whole try/catch/.../catch
sequence as a single statement. An alternative would be to process them
as separate statement that we logically link after the fact, like
let/and/.../and sequences are).