bootstrap:passbody: literal conversion call inference: skip on ungenarg DEFINTF
The existing code doesn't work when trying to use typ_member(t,
ID_FROM_STRING) when t is a DEFINTF and the method ID_FROM_STRING (say)
is defined not in t but in an "inherited" interface 'i'.

Indeed, 'final' in the member is not evaluating to 't', but evaluating
to 'i' (because of the way scope resolution and inference works...).

Anyway, when t is a DEFINTF and an ungenarg, we don't need to do this
transform (as the code isn't concrete), so we just skip the
transformation.

Not sure about the case where t is a DEFINTF and not an ungenarg. It
should be because of a dyn and we should be seeing @t, not t. Maybe?