bootstrap/ncc0: support for emcc; env NCC_COMPILER to select compiler
Like for clang, support is not perfect (clang generates a couple of annoying
warnings, repeated many times, on the code emitted by cprinter).

But several tests have been checked to pass. trun doesn't have support
for emcc yet (it's mostly a matter of getting rid of the clang warnings,
actually).