bootstrap: more flexible try-less except: can return (Error, ...)
A fundamental idiom in the language -- to be enforced by semantic
constraints -- says that in a tuple returned by a function, any value
that appears after an Error is (not ^Init) if the error is not OK.

So we can sensibly support this:

	fun foo = Error, U32
		except = ...

Equivalent to
	fun foo = Error, U32
		let err = ...
		if err != OK
			return err, {}

However, the following is rejected:

	fun foo = U32, Error
		except = ...

as we don't know what U32 value to return, and it's not acceptable to
assume it's 0.

By making except more flexible in this manner, we're encouraging the
application of this idiom.

We could even accept to do this if the return type is (Error,
No_default_ctor), but the compiler would currently reject the
transformed LIR, so we reject for the moment.