bootstrap:types: olay: map children of generic functor types
For instances, we were properly mapping children (map_children). But for
(uninstantiated) generic functor, map_children was called too early, as
the typ is built before the children (member functions) are themselves
typed. So in that case, we must do another pass after the typing of
children, to map them in the parent typ.

The explosion in the number of cases that must be mapped, made even more
apparent with this change, means we will have to implement on-demand
filed overlays.