bootstrap:inference: on-demand typing of aliases: allow more definition orders
Aliases and isalists are typed in the same passfwd, allowing for one to
depend on the other in a single module, in either order (isa an_alias or
alias a = {`literal `list}), according to the declaration order in the
file.

With this change, we allow more flexibility, typing a DEFALIAS on demand
if it's found as part of a bin_acc lookup (e.g. io.`Writer). This is a
stopgap solution to the compiler using a pure dependency order.