bootstrap: support named examples with one bench.Example argument
	example foo ex:@!bench.Example
	        ex!Benchmark FS
		...
		assert ex.Fs.Opened == 1

Basic environments are now equipped with built-in statistics (e.g. Fs,
Heap, ...). These statistics are locally resettable, they form a stack.

	ex!Bencharm FS HEAP

resets (and could enable -- they're enabled by default now) the statics
for the environments Fs and Heap. So they start at zero at the beginning
of the example. Numbers can be checked at the end of the example. The
local stats are popped from the stack at the end of the example, and
incorporated into the previous stats.

This last feature is very useful when debugging.