bootstrap:cprinter: experiments with always_inline unconclusive
Using always_inline doesn't seem to be hugely beneficial, although there
appears to be a small improvement in runtime (<= 5%). As expected, the
nature of the generated code (accessors, many thin methods, etc) may
demand that we tweak inlining rules in the C compilers. But it's
reassuring that they seem to do a pretty good job on their own.