bootstrap: propagate local static constants, such as generated Codeloc
Turns out that C compiler will not optimize

	struct Codeloc c = { 0 };
	c.Line = 42;
	Assert__(cond, &c);

into, after inlining:

	if (cond) {
	  struct Codeloc c = { .Line = 42 }:
	  // print 'c' then abort
	}

Instead, we get something like:

	struct Codeloc c = { 0 };
	c.Line = 42;
	if (cond) {
	  // print 'c' then abort
	}

The cost of the copy that initializes 'c' is paid even if cond is false,
even though 'c' is not used outside of that branch. It appears the
compilers (gcc and clang) refuse to move the initialization of 'c' into
the only branch where it's used. Maybe they're not allowed to, but I
can't see why.

Instead, we generate the following code:

	static struct Codeloc c = { .Line = 42 };
	Assert__(cond, &c);

which optimizes nicely, as the initialization of 'c' is moved out of the
way entirely.

We reuse the global constant propagation logic, using a new flag
NODE_IS_LOCAL_STATIC_CONSTANT. It must used with care, as the wrong use
of 'static' would have dire consequences. It is currently applied only
to STRING nodes (which are truly immutable) and generated _Ncodeloc
arguments.

In general, we cannot use it on

	DEFNAME
	 IDENT(x)
	 NUMBER

as the location of 'x' may be modified.