bootstrap:printer: debug global ppptr: if set, pptree() prints node addresses
This is very useful when trying to break when a node is being processed.
Say you're trying to understand why a node is typed a certain way.

	$ run
	Ctrl-c early
	$ set ppptr = 1
	$ continue

Check the output in t00/file.n.o.tree (or call pptree() directly). Take
the address of the node, e.g.:

	0x00000001	LET :n.builtins.Void
	0x000000f0	 DEFNAME(_Ngensym1) :*(n.builtins.`Any_ref n.builtins.I32)

	$ watch -l ((struct node *) 0x000000f0)->typ
	$ run

gdb will break whenever 'typ' is updated.