bootstrap:types: on-demand overlays 4/4: translate: find existing overlay or create one
Previously, we would create overlays for all possible dependencies of a
generic: all children, all isas, all direct or indirect users of its
ungenarg; and all that recursively. That was of course crazy and
explosive. Especially because so many tentative generic instances are
very transient and that see barely any use before being linked away.

In other words, we were doing a lot of work, "just in case" it was
needed, even though the great majority of the time, it isn't.
Compilation times had recently ballooned out of control, as we added
more and more obscure corner cases to the "just in case" list.

The proper approach requires subtlety, but isn't as hard as we may have
feared. On 'olay(t, u)', we determine if 'u' needs to have an overlay
with respect to 't' (e.g. 'u' shares ungenargs with the functor of 't'),
and we create the required overlay on the fly.

The compilation is much faster, close to reasonable, pointing to a
bottleneck in typ_isa.