1/2 Revert "bootstrap:types: do not duplicate a functor quickisa table in its instances quickisa"
Turns out this wasn't a good idea, at least as-is. It was done to keep
quickisa tables smaller and to avoid having to update the quickisa of
users of a generic functor when the quickisa of the functor itself is
updated (e.g. on autointf).

As written, this patch was wrong.

This reverts commit 0ffe40d37f740e479805109e6ef6738d11f66101.

Conflicts:
	bootstrap/types.c