Makefile: bootstrap: use -iquote not -I to point to lib/
Using -I, it is possible to use #include <lib/...>, as was done in
reflect.h. The problem is that gcc then may treat files that include
reflect.h (here nodes.h) as a system header, under certain
circumstances. Namely, because of a ccache recompilation avoidance trick
(check the cpp output for differences before really deciding to
recompiler), gcc is passed that cpp output directly. And somehow,
nodes.h code ends up with such a line directive:

	# 405 "bootstrap/nodes.c" 3 4

3 is a flag that indicates it should be treated as a system header, in
which most warnings are silenced.

According to the doc of gcc, this should only happen for headers found
in -isystem. It is not at all clear why nodes.h is seen as a system
header. This appears to be a gcc bug.

In any case, there is a different gcc bug in that this problem does not
appear if

	export CCACHE_RUN_SECOND_CPP

is set. Then ccache doesn't feed gcc the cpp output, but let gcc do the
whole chain itself. And this works just fine. gcc shouldn't behave
differently with either technique.

We do set CCACHE_RUN_SECOND_CPP to limit our exposure to such
differences in behavior in gcc, sticking to what is likely the most
common way to use gcc.

	gcc (GCC) 4.9.2