bootstrap: automagic: implement deopt, deref, opt (no automagic ref)
No automagic ref when initializing a struct is a language design
choice. It is not entirely unreasonable to allow it, but without escape
analysis, it's not very smart: unlike in the case of function arguments
-- functions don't escape their argument that often, INIT expressions
are very often escaped...

Some of this (deopt/opt) was done in cprinter; it's now done in
passbody1, where it belongs.