n.builtins.Hashset: support and `Set interface
Enumset can be `Universal_set, but Hashset cannot.

As ~ cannot really be defined on Hashset (because, while we could define
something like ALL & ~{x, y, ...}, we cannot iterate over a set that
does not contain some elements).

So we define `Set, a narrower set of operations not based on bitwise
operators.