bootstrap:cprinter: cleanup non `Trivial_dtor local variables
Rely on the __cleanup__ attribute (gcc and clang extension).

Introduce the temporary concept "is_stolen_by_move" for variables that
serve as Copy_ctor destination temporaries, that are then moved to
another memory location, such as a TUPLE or INIT. But this is
ill-defined, it only makes sense for `Trivial_move types, and we are
entirely unable to identify who they are.

FIXME(e): when not `Trivial_copy, cannot use literal assignment
expressions in the generated C code, and get rid of is_stolen_by_move.