bootstrap:lir: MATCH becomes a HIR-only construct, rewrite into IF
While there may be value for codegen and the semantic constraint passes
to preserver the MATCH structure, it does add one more case to handle in
many places.

We simply rewrite MATCH into a IF tree, using the TEQMATCH operator.

The TEQMATCH and TNEMATCH are now rewritten into TEQ and TNE in
inference.c once the type of the arguments are known, rewriting their
terms into tag accessors when necessary. This used to be done in
cprinter, by hand, but this isn't acceptable anymore to introduce such
computations that late in the process, now that the compiler generates
Copy_ctor and Dtor calls.