bootstrap:inference: {}:union.A case wasn't unified properly
In unify_with_defincomplete_entrails we don't yet know the tag (e.g. A)
so we merely checked whether the field names were possible in the target
type. That's fine.

But in type_inference_typeconstraint_defchoice_init, we were merely
linking the types of the values in the INIT with the target fields
types; we should have been calling unify. We do so after calling
try_insert_automagic_de.

We can then skip this case step_init_insert_automagic, in passbody1.