n.builtins: use `Fmt_state in Show prototypes, not fmt.`State
This makes it easier to use newtype on builtin types: there is no need
to explicitly import fmt.

This is because newtype copies prototype literally, which means it is
assumed that the environment where the newtype is defined is similar
enough to n.builtins. n.builtins imports fmt, but not everyone does.