builtins.Map: Try_store: return pair instead of nullable ref
Try_store returned nil if the key was already in the map, so the
caller wouldn't overwrite an existing value. But to get the existing
value, the caller had to do a second call.

Instead, Try_store now returns a pair (Bool, #T): the first element
indicates if a new location had to be created, and the second element is
a reference either to the new location or to the existing value.