bootstrap:useorder: build C-like dependency graph
This is a first step. We build a dual topdep structure, topdeps_td, to
be used by useorder. This is just to keep things simple, these 2
structures are likely fully redundant.

For each pass of cprinter, useorder is capable to extract a suitable
node ordering from topdeps_td.

In the future, topdeps will not be used for stage (compilation)
ordering, we will instead maintain a (partial) order over the all stage,
updating it as we discover new dependencies and their exact nature.