bootstrap:autointf: generate `Any.Sizeof, Alignof, and Move
The Move method is generated by the compiler, as a simple C copy and
zeroing of the original memory. The value of self is copied as a return
value.

Move support in the compiler is nearly magic.

When performing a move, the compiler must not generate a spurious copy
constructor call. We have two cases:

- `Return_by_copy types are all `Trivial_copy. So no problem there.

- The other types are return-through-ref. 4 cases:

1. Assignment

	x = y#Move

becomes in C:

	type_y$Move(&y, &x);

The burden of any copy is relegated to the function being called. So no
problem here either.

2. Return value:

	return x#Foo

becomes in C (again, return-through-ref!):

	type_retval$Move(&_nretval, &x);

and we're good.

3. Argument passing:

4. Init assignment: