n.builtins.Buf: require t:{`Moveable `Trivial_ctor `Default_dtor}
There are still some questions here. A large subset of Buf doesn't
depend on these operations: even Push can be performed on a
non-`Trivial_ctor type (at least with the help of ::init tracking by the
semantic engine). We don't want to push people away from Buf because of
stringent requirements, to have them reimplement it by hand and get it
wrong.