n.builtins.Map: sptable: use (Buf (Buf T)) instead of (Buf ?#(Buf T))
The level of indirection was costly for very large Maps.

In essence, hashtables have random memory access patterns, necessarily
non-local on subsequent set/get operations. For large hashtables, the
additional level of indirection resulted in a second cache fault.

Now that Buf is `Trivial_move, we can avoid the indirection. The
performance of Map is now with 5% of bootstrap/table.h, for pretty much
any size.