bootstrap: Move constructor and semantics
Function arguments are now owned by the callee, and it is responsible to
Dtor them if needed.

This allows the callee to Move its argument back out to the caller.

And this also allows the caller to Move argument values down to the
callee, which will Dtor them (or, as we just said, Move them back).

t00/move covers a lot of cases, but it is probably not quite complete.

Valgrind reports no leaks so far (not all tests were run that way).