bootstrap:ssa: defname_replace_block_like_expr: prevents removal if not void
A basic ssa_sub() creates a single "ssa_user" for the new defname. So it
is easy to remove it if it turns out, after typing, that the sub was
unnecessary (e.g. it was a field accessor expression, etc.).

But with defname_replace_block_like_expr() we create two users for the
defname, and we don't know how to remove such unnecessary subs (yet).
Unless they're void.

FIXME: we need to learn to remove such unnecessary subs in this case
too, as there are cases where such subs are illegal (illegal naive
copy, bad reference).