Change `Iterator interface: single Next method returning an opt value
The Has_next/Next combination, while traditional, has many problems. For
instance, writing a filtering iterator wrapper is challenging: Has_next
must look up ahead in the underlying iterator to decide if the available
elements shall pass the filter or not. This cannot be done without
mutating the underlying iterator and having state. Not ideal when
Has_next is supposed to be a const method.

This actually led us to write a buggy Map_iter: it acts as a filtering
iterator on top of sptable_iter.

We are most definitely not advocating that everything be given a
standard iterator interface: it is rarely the right tool for the job.
However, if a number of types are going to offer one in the N standard
library, we should still do our best to choose an interface that is not
error prone, and is as sane as possible.